+++ components/core/footer.jsx
